TUTORIAL DONE

+++ app/Http/Controllers/ListingController.php
@@ -67,7 +67,7 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing){
         // check if user is the owner of the listing
         if($listing->user_id !== auth()->user()->id){
-            return redirect('/')->with('error', 'Unauthorized action');
+            abort(403,'Unauthorized action');
         }
         $validatedData = $request->validate([
             'title' => 'required|max:255',
@@ -100,6 +100,10 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+        // check if user is the owner of the listing
+        if($listing->user_id !== auth()->user()->id){
+            abort(403,'Unauthorized action');
+        }
         //dd($listing);
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
